requirements.txt
pymyq>=3.1.6
gspread>=5.12.0
PyYAML>=6.0
aiohttp>=3.9.0
